test: add more tests, linting
